requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.2
pydantic>=2.11.1
python-dotenv>=1.1.0
requests>=2.32.3
openai>=1.76.0
torch>=2.3.0
Pillow>=11.2.1
yfinance>=0.2.56
python-multipart>=0.0.20
diffusers>=0.33.1
transformers>=4.51.3
accelerate>=1.6.0
beautifulsoup4>=4.13.4
pymupdf>=1.25.1
docx2txt>=0.8
pytesseract>=0.3.13
langchain_community>=0.3.19
langchain>=0.3.20
langchain_chroma>=0.2.2
chromadb>=0.6.3
sentence_transformers>=4.1.0
langchain_huggingface>=0.1.2